third_party/blink/web_tests/external/wpt/css/css-anchor-position/parsing/position-try-options-computed.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Anchor Positioning Test: Computed position-try-options</title>
<link rel="help" href="https://drafts.csswg.org/css-anchor-position-1/#position-try-options">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
<div id="container">
  <div id="target"></div>
</div>
<script>
  test_computed_value("position-try-options", "none");
  test_computed_value("position-try-options", "flip-block");
  test_computed_value("position-try-options", "flip-inline");
  test_computed_value("position-try-options", "flip-start");
  test_computed_value("position-try-options", "flip-block, flip-inline");
  test_computed_value("position-try-options", "--foo, --bar");
  test_computed_value("position-try-options", "flip-start flip-inline flip-block");
  test_computed_value("position-try-options", "flip-start --flop", "--flop flip-start");
  test_computed_value("position-try-options", "--flop flip-start");
  test_computed_value("position-try-options", "left top");
  test_computed_value("position-try-options", "top left", "left top");
  test_computed_value("position-try-options", "start start", "start");
  test_computed_value("position-try-options", "left, right");
  test_computed_value("position-try-options", "--foo, left");
  test_computed_value("position-try-options", "--foo, left, --bar");
  test_computed_value("position-try-options", "--foo, flip-start, left");
  test_computed_value("position-try-options", "--foo flip-start, left");
  test_computed_value("position-try-options", "left, --bar flip-start");

  assert_not_inherited("position-try-options", "none", "flip-inline");
</script>
